recovery: Move RunResult to proto
BUG=b:233690127
TEST=build

Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3662447
Commit-Queue: Otabek Kasimov <otabek@google.com>
Auto-Submit: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#49271}

+++ go/src/infra/cros/recovery/internal/localtlw/ssh/ssh.go
@@ -90,7 +90,7 @@ func createSessionAndExecute(ctx context.Context, cmd string, client *ssh.Client
 		case nil:
 			result.ExitCode = 0
 		case *ssh.ExitError:
-			result.ExitCode = t.ExitStatus()
+			result.ExitCode = int32(t.ExitStatus())
 		case *ssh.ExitMissingError:
 			result.ExitCode = -2
 			result.Stderr = t.Error()
